Bump to v1.0.0 (#83)
* Use fresnel propagation by default (#71)

* Use fresnel propagation by default

* Fix propagation method checks in `is_angular_spectrum_method`

* Float-default (#72)

* use float by default

* Update modules to use float by default

* Fix complex behaviour

* Add uv (#74)

* Fixes (#75)

* Move torchoptics package into src layout

* Refactor pyproject.toml to use hatchling for build system and update project metadata

* change name to types.py

* use collections

* Update Python version requirement to 3.10 and remove 3.9 from classifiers

* Remove typing Optional and Union

* Propagation gradient fix (#76)

* Refactor asm padding arg and copy method (#77)

* Make copy a Field method

* Refactor propagation methods to use keyword-only arguments for clarity

* Refactor ASM padding parameter from asm_pad_factor to asm_pad for consistency across methods

* Add dedicated tests for Field.copy() (#78)

* Fix ruff linter errors (#79)

* Make copy a Field method

* Refactor propagation methods to use keyword-only arguments for clarity

* Refactor ASM padding parameter from asm_pad_factor to asm_pad for consistency across methods

* Ruff fixes

* Fix ruff errors

* Update src/torchoptics/profiles/hermite_gaussian.py

* Fix docstring types for `create_image_subplot()` to match type annotations (#80)

* Update GitHub Actions workflows and improve pre-commit configuration

* Update docs

* Update .gitignore to include examples directory for Sphinx gallery

* Add __version__

* Update Python version badge to 3.10+

* Remove copyright

* Update normalization tolerance in laguerre gaussian tests

* Add tests for gradient propagation through PlanarGrid in propagate_to_plane

+++ tests/system/test_system.py
@@ -14,9 +14,9 @@ def make_system_setup():
     modulator1_offset = (0.0, 0.0)
     modulator2_spacing = 6e-6
     modulator2_offset = (1.3e-6, -3.5e-6)
-    square_field = torch.ones(shape, shape, dtype=torch.cdouble)
+    square_field = torch.ones(shape, shape, dtype=torch.cfloat)
     input_field = Field(square_field, spacing=spacing, wavelength=wavelength)
-    mod_profile = torch.ones(shape, shape, dtype=torch.cdouble)
+    mod_profile = torch.ones(shape, shape, dtype=torch.cfloat)
     return (
         shape,
         spacing,
